Reorganize the different SQL vs Elasticsearch log classes.
This shifts things around so that we're not just hard-coding everything
to the duplicate SQL classes. Instead, there's now a configurable
concept of the analytics storage adapter, and these classes now inherit
from a base class, so there's less duplication.
