+++ src/api-umbrella/web-app/app/controllers/api/v1/analytics_controller.rb
@@ -5,7 +5,7 @@ class Api::V1::AnalyticsController < Api::V1::BaseController
   around_filter :set_time_zone
 
   def drilldown
-    @search = LogSearchSql.new({
+    @search = LogSearch.factory({
       :start_time => params[:start_at],
       :end_time => params[:end_at],
       :interval => params[:interval],
