+++ src/api-umbrella/web-app/app/models/log_result/elasticsearch.rb
@@ -1,0 +1,2 @@
+class LogResult::Elasticsearch < LogResult::Base
+end
