+++ src/api-umbrella/web-app/app/models/log_result/kylin.rb
@@ -1,0 +1,2 @@
+class LogResult::Kylin < LogResult::Sql
+end
