+++ src/api-umbrella/web-app/app/models/log_result/postgresql.rb
@@ -1,0 +1,2 @@
+class LogResult::Postgresql < LogResult::Sql
+end
