+++ src/api-umbrella/web-app/app/models/log_search/kylin.rb
@@ -1,0 +1,2 @@
+class LogSearch::Kylin < LogSearch::Sql
+end
