+++ src/api-umbrella/web-app/app/models/log_search/postgresql.rb
@@ -1,0 +1,2 @@
+class LogSearch::Postgresql < LogSearch::Sql
+end
